# Remove special-cases for scalars in contractions