It's possible this will be marginally slower, but it makes everything much easier to read.