@eric-wieser 2019-12-02
real failure:
```
    def test_check_generalized_BAC_CAB_formulas(self):
    
        (a,b,c,d,e) = Ga('a b c d e').mv()
    
>       assert str(a|(b*c)) == '-(a.c)*b + (a.b)*c'
E       AssertionError: assert '(b.c)*a - (a.c)*b + (a.b)*c' == '-(a.c)*b + (a.b)*c'
E         - (b.c)*a - (a.c)*b + (a.b)*c
E         ? -------- -
E         + -(a.c)*b + (a.b)*c
```