@eric-wieser 2019-12-04
I need to investigate this some more. Some of the failures are just addition reordering, which can be easily fixed.

The other half look like actual bugs. Leaving this as a draft for now, I probably need a piecewise approach to work out which bit is broken.