@eric-wieser 2019-12-06
Some timings:
```
In [2]: from galgebra.ga import Ga

In [3]: g, *a = Ga.build('e*1|2|3|4')

In [4]: A, B = g.mv('A', 'mv'), g.mv('B', 'mv')
```
Before:
```
In [5]: %timeit A < B
102 ms ± 2.34 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)
```
After:
```
In [7]: %timeit A < B
105 ms ± 1.91 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)
```

No significant change in the general case. Probably faster for a pure scalar contraction, but that doesn't seem like a case worth optimizing for.